AVX optimized implementation of resize and warp functions migrated to separate file
